% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{bboxm3}
\alias{bboxm3}
\title{Límites (\code{bbox}) para los Mapas en MEDEA3}
\format{Una lista donde el primer elemento es el código de municipio INE, el
  segundo es una lista con los bbox por ciudad y el tercero un vector
  definiendo la posición de la leyenda.}
\usage{
bboxm3
}
\description{
Lista con los \code{\link[sp]{bbox}} para las ciudades MEDEA3.
}
\examples{

\dontrun{
  library(medear)
  data("bboxm3")
}
}
\keyword{datasets}
