% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cambios_pais}
\alias{cambios_pais}
\title{Cambios de seccionado para todo el país.}
\format{Un objeto de clase \code{cambios_ine}, donde cada
  fila es un un cambio de sección y que cuenta con 4 columnas:
  \describe{
    \item{sc_old}{Cádena de 10 caracteres con el código de la sección en año
    == year.}
    \item{sc_new}{Cádena de 10 caracteres con el código de la sección en año
    == year2.}
    \item{year}{Primer año.}
    \item{year2}{Segundo año.}
  }}
\usage{
cambios_pais
}
\description{
Cambios de seccionado para todo el país.
}
\examples{

\dontrun{
library(medear)
data(cambios_pais)
}
}
\keyword{datasets}
