% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{carga_datos}
\alias{carga_datos}
\title{Carga los datos privados de población para el proyecto MEDEA3}
\usage{
carga_datos(key)
}
\arguments{
\item{key}{Cadena de caracteres con la contraseña.}
}
\value{
No se devuelve nada.
}
\description{
Algunos datos del proyecto MEDEA3 están encriptados para poder
  cumplir con la licencia INE (poblaciones desde 1998 a 2003). Esta función
  los desencripta y los adjunta al entorno global.

  CUIDADO: se sobreescribirá cualquier objeto nombrado como poblacion.
}
\details{
La contraseña no se almacena en el historial.
}
\examples{
\dontrun{
  carga_datos(key = "contraseña")
}

}
\seealso{
poblacion
}
\keyword{datasets}
