% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{carto_medea3}
\alias{carto_medea3}
\title{Cartografía por Sección Censal para las Ciudades MEDEA3 (1996-2015)}
\format{Un objeto de clase \code{SpatialPoligonsDataFrame}, donde los datos
  (\code{carto_medea3@data}) tienen clase \code{data.frame}, donde cada fila
  es una sección censal y que cuenta con 3 columnas: \describe{
  \item{seccion}{Cádena de 10 caracteres con el código de sección censal
  (incluye provincia, municipio, distrito y sección).} \item{CUMUN}{Cádena de
  5 caracteres con el código INE del municipio.} \item{NMUN}{Nombre del
  municipio.}}}
\usage{
carto_medea3
}
\description{
Contiene la cartografía de secciones censales unificada para el
  proyecto MEDEA3 (período 1996-2015), tomando como cartografía de referencia
  la proporcionada
  \href{http://www.ine.es/censos2011_datos/cen11_datos_resultados_seccen.htm}{por
  el INE para el año 2011}.
}
\examples{

\dontrun{
  library(medear)
  library(sp)
  data("carto_medea3")

  # Representación de los secciones censales de Álava
  plot(carto_medea3[substr(carto_medea3$seccion, 1, 5) == "01059", ])

  # Representación de los secciones censales de Álava, según distritos.
  distritos <- substr(carto_medea3[substr(carto_medea3$CUSEC, 1, 5) == "01059", "CUSEC"], 6, 7)
  plot(carto_medea3[substr(carto_medea3$CUSEC, 1, 5) == "01059", ], col = as.numeric(distritos))
}
}
\references{
\href{http://www.ine.es/}{Sitio web del INE}.

  \href{http://www.ine.es/censos2011_datos/cen11_datos_resultados_seccen.htm}{Cartografía}.
}
\keyword{datasets}
