% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cartografia}
\alias{cartografia}
\title{Cartografía por sección censal para las ciudades MEDEA3}
\format{Un objeto de clase \code{cartografia_ine} y \code{sf}, donde cada
  fila es una sección censal y que cuenta con 7 columnas: \describe{
  \item{seccion}{Cadena de 10 caracteres con el código de sección censal
  (incluye provincia, municipio y distrito).} \item{cumun}{Cadena de 5
  caracteres con el código del municipio (incluye provincia).}
  \item{ccaa}{Cadena de 2 caracteres con el código de comunidad autónoma.}
  \item{npro}{Nombre de la provincia.} \item{nca}{Nombre de la comunidad
  autónoma.} \item{nmun}{Nombre del municipio.} \item{geometry}{Columna de
  tipo lista con la geometría asociada a cada sección censal.} }}
\usage{
cartografia
}
\description{
Contiene la cartografía por sección censal en formato Simple
  Features tal cual puede ser utilizada por el paquete \code{sf}.
}
\examples{

\dontrun{
  library(medear)
  data("cartografia")

  # Representación de las secciones censales de Álava
  plot(cartografia[cartografia$cumun == "01059", "seccion"])
}
}
\references{
\href{http://www.ine.es/}{Sitio web del INE}.
\href{http://www.ine.es/censos2011_datos/cen11_datos_resultados_seccen.htm}{Cartografía}.
}
\keyword{datasets}
