% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{censo}
\alias{censo}
\title{Censos de 2001 y 2011}
\format{Un objeto de clase \code{data.frame} con los siguientes campos:

  \itemize{
    \item seccion Código de sección censal (10 caracteres).
    \item muni Código INE del municipio.
    \item year Año del censo.
    \item i01 Número de trabajadores manuales entre los ocupados o parados que
      han trabajado antes, de 16 o más años.
    \item i02 Número de trabajadores manuales entre los ocupados, de 16 o más años.
    \item i03 Número de parados de 16 o más años entre los activos.
    \item i04 Número de asalariados eventuales entre los ocupados o parados
      que han trabajado antes de 16 o más años.
    \item i05 Número de asalariados eventuales entre los ocupados de 16 o más años.
    \item i06 Número de personas con instrucción insuficiente de 16 o más años.
    \item i07 Número de personas con instrucción insuficiente de 16 a 29 años.
    \item i08 Número de población de 65 o más años.
    \item i09 Número de nacidos en países de renta baja.
    \item i10 Número de nacidos en países de renta baja llegados a España
      después de 2006.
    \item i11 Número de nacidos en países de renta baja o nacidos en España
      cuyo padre o madre ha nacido en países de renta baja.
    \item IE01_d Total de población ocupada o parada que ha trabajado antes
      de 16 o más años.
    \item IE02_d Total de población ocupada de 16 o más años.
    \item IE03_d Total de población de 16 o más años en situación laboral = 1
      Ocupado/a + 2 Parado que ha trabajado antes + 3 Parado que busca
      empleo (población activa).
    \item IE06_d Total de población de 16 o más años.
    \item IE07_d Total de población de 16 a 29  años.
    \item IE08_d Total de población total.
  }}
\description{
Datos de los censos de 2001 y 2011 necesarios para construir el
  índice de privación. Estos datos se cargan con la función
  \code{\link{carga_datos}}.
}
\examples{

\dontrun{
  library(medear)
  censo <- carga_datos("contraseña", tipo = "censo")
}

}
\seealso{
\code{\link{carga_datos}}
}
\keyword{datasets}
