% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_ine}
\alias{codigos_ine}
\title{Código de provincias y municipios del INE}
\format{Un objeto de clase \code{data.frame}, donde cada fila es un municipio
  y que cuenta con 7 columnas: \describe{ \item{cod_ccaa}{Cadena de 2
  caracteres con el código de la comunidad autónoma.}
  \item{cod_provincia}{Cadena de 2 caracteres con el código de la provincia.}
  \item{cod_municipio}{Cadena de 3 caracteres con el código del municipio.}
  \item{nombre_ccaa}{Nombre de la comunidad autónoma.}
  \item{nombre_provincia}{Nombre de la provincia}
  \item{nombre_municipio}{Nombre del municipio.} \item{medea3}{Vector lógico:
  ¿participa en MEDEA3?} }}
\usage{
codigos_ine
}
\description{
Codificación usada por el INE para las provincias y los
  municipios. También indica si se trata de un municipio participante en
  MEDEA3 o no.
}
\examples{

\dontrun{
  library(medear)
  data("codigos_ine")
}
}
\references{
\href{http://www.ine.es/}{Sitio web del INE}.
\href{http://www.ine.es/daco/daco42/codmun/codmunmapa.htm}{Codificación INE}.
}
\keyword{datasets}
