% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{comprueba_datos}
\alias{comprueba_datos}
\title{Comprobaciones de la clase de los datos}
\usage{
comprueba_datos(datos, tipo = c("cartografia", "mortalidad",
"poblacion"), periodo = NULL)
}
\arguments{
\item{datos}{Datos sobre los que lanzar comprobaciones.}

\item{tipo}{Tipo de datos a comprobar.}

\item{periodo}{Período de estudio: solo aplicable en \code{tipo == "mortalidad"}
o \code{tipo == "poblacion"}.}
}
\description{
Función para comprobar si los datos se ajustan al formato requerido.
}
