% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{comprueba_geocodificado}
\alias{comprueba_geocodificado}
\title{Comprobación de la asignación de distintas coordenadas a una misma
  dirección}
\usage{
comprueba_geocodificado(mortalidad)
}
\arguments{
\item{mortalidad}{Datos con la mortalidad geocodificada. Objeto de clase
\code{data.frame} que contenga, al menos, la siguiente información (los
nombres que deben tener las variables están entre paréntesis): longitud
(\code{lng}), latitud (\code{lat}), tipo de vía (\code{tip_via}),
dirección (\code{address}), número de portal (\code{portalNumber}),
municipio (\code{muni}) y provincia (\code{province}) devueltos por el
servicio de geocodificado.}
}
\value{
Devuelve un \code{data.frame} con la mortalidad y los registros
  problemáticos modificados a \code{NA}, fijando el campo \code{georef} de
  esos registros al valor \code{repetir_geo_google}.
}
\description{
Aunque no es frecuente, en algunas situaciones, y frente a una
  misma dirección, se obtienen distintas coordenadas. Esta función detecta
  estos casos para que pueda volver a lanzarse el algoritmo de geocodificado
  sobre ellos. Concretamente se modifican estos registros de los datos de
  mortalidad, sustituyendo los valores devueltos por el servicio de
  geocodificado por \code{NA}'s, y fijando
  \code{georef == "repetir_geo_google"}.
}
