% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crea_cubo_mortalidad}
\alias{crea_cubo_mortalidad}
\title{Crear la matriz 5-dimensional de mortalidad}
\usage{
crea_cubo_mortalidad(datos, cartografia, epsg = 4326, medea3 = TRUE,
  otras_causas = NULL, corte_edad = 85, periodo = 1996:2015)
}
\arguments{
\item{datos}{ll}

\item{cartografia}{ll}

\item{epsg}{ll}

\item{medea3}{ll}

\item{otras_causas}{Véctor de caracteres que indica los nombres de las
columnas (columnas con valor 0-1) en la base de datos de mortalidad que
identifican a dichas otras causas.}

\item{corte_edad}{ll}

\item{periodo}{ll}
}
\description{
lala
}
