% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crea_cubo_poblacion}
\alias{crea_cubo_poblacion}
\title{Crear la matriz 4-dimensional de poblaciones}
\usage{
crea_cubo_poblacion(datos, cartografia, epsg, periodo = 1996:2015,
  datos_propios = FALSE)
}
\arguments{
\item{datos}{ll}

\item{cartografia}{ll}

\item{epsg}{ll}

\item{periodo}{ll}

\item{datos_propios}{ll}
}
\description{
lala
}
