% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{data_vignette}
\alias{data_vignette}
\title{Datos para construir las viñetas MEDEA3}
\format{An object of class \code{list} of length 2.}
\usage{
data_vignette
}
\description{
Lista con el índice de privación e I de Moran
  del mismo para las ciudades MEDEA3 (siguiendo cartografía unificada MEDEA3).
}
\examples{

\dontrun{
  library(medear)
  data("data_vignette")
}
}
\keyword{datasets}
