% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\encoding{UTF-8}
\name{descarga_cartografia}
\alias{descarga_cartografia}
\title{Descarga la cartografia con el seccionado del INE para 2011}
\usage{
descarga_cartografia(epsg = 4326, conservar = TRUE, ntries = 10)
}
\arguments{
\item{epsg}{Vector numérico de longitud uno con el código EPSG del sistema de
referencia de coordenadas (CRS) empleado (por defecto se usa el 4326 con
datum WGS84).}

\item{conservar}{Valor lógico: ¿se desea conservar los archivos descargados
en el directorio oculto \code{./.cartografia/} dentro del directorio de
trabajo?}

\item{ntries}{Valor numérico: número de intentos en caso de mala conexión.}
}
\value{
Un objeto de clase \code{\link[sp]{SpatialPolygons}}, donde
  cada fila es una sección censal y que cuenta con 7 columnas:
  \item{seccion}{Cadena de 10 carácteres con el código de sección censal
  (incluye provincia, municipio y distrito).} \item{CUMUN}{Cadena de 5
  carácteres con el código del municipio (incluye provincia).}
  \item{CCA}{Cadena de 2 carácteres con el código de comunidad autónoma.}
  \item{NPRO}{Nombre de la provincia.} \item{NCA}{Nombre de la comunidad
  autónoma.} \item{NMUN}{Nombre del municipio.}
}
\description{
Descarga la cartografía del seccionado censal ofrecida
  públicamente por el INE para el año 2011.
}
\details{
Aunque el INE emplea otro CRS, se recomienda utlizar el CRS 4326
  como elemento normalizado.
}
\examples{

\dontrun{
  library(medear)
  library(sp)
  carto_ine    <- descarga_cartografia()
  carto_ine_46 <- carto[substr(carto$seccion, 3, 4) == "46", ]
  plot(carto_ine_valencia)
}

}
\seealso{
\code{\link{descarga_trameros}} y
  \code{\link{descarga_poblaciones}}.
}
