% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\encoding{UTF-8}
\name{descarga_poblaciones}
\alias{descarga_poblaciones}
\title{Carga poblaciones del INE por sección censal, sexo, edad y año}
\usage{
descarga_poblaciones(cod_provincia = c(paste0("0", 1:9), 10:52), years =
  2006:2016, descarga = TRUE, ruta = NULL)
}
\arguments{
\item{cod_provincia}{Cadena de caracteres de longitud >= 1 con el código de
la/s provincia/s en las que se desee obtener el listado de cambios de
seccionado.}

\item{years}{Vector numérico de longitud >= 1 con los años para los que se
desee consultar las variaciones de seccionado.}

\item{descarga}{Valor lógico: ¿debe procederse a la descarga de los datos?}

\item{ruta}{Cadena de caracteres indicando la ruta en la que se almacenan los
archivos tal cual se descargaron desde el INE, en caso de escoger
\code{descarga = FALSE}.}
}
\value{
Un objeto de clase \code{poblaciones_ine} donde las filas representan
  las distintas secciones censales. Las tres primeras columnas son:
  \item{seccion}{Código de la sección censal en el primer año.}
  \item{sexo}{Código de la sección censal en el segundo año.}
  \item{year}{Primer año.}
  El resto de columnas representan los distintos grupos de edad.
}
\description{
Descarga o carga las poblaciones anuales del INE por sección
  censal, sexo y edad por grupos quinquenales (datos desde 2006).
}
\details{
El tiempo de ejecución de la función varía según el número de
  provincias y el rango de años. La forma más sencilla de acelerar el proceso
  de computación es mediante la ejecución en paralelo de la función.

  Los códigos de sección censal siguen un orden preestablecido: los primeros
  dos dígitos identifican la provincia, los siguientes tres dígitos el
  municipio, los próximos dos dígitos el distrito y los últimos cuatro
  a la sección censal.

  Hasta el año 2011 el INE agrupa la última categoría de edad como 85 y más,
  mientras que desde el año siguiente llega hasta 100 y más.
}
\examples{

\dontrun{
poblaciones <- descarga_poblaciones(cod_provincia = "46")
poblaciones
}

}
