% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{detecta_cluster}
\alias{detecta_cluster}
\title{Deteccion de agrupaciones de mortalidad a revisar manualmente}
\usage{
detecta_cluster(datos, epsg = 4326, columnas = c("lng", "lat"),
  vecinos = 10, limite = 1e-20)
}
\arguments{
\item{datos}{Base de datos con las coordenadas que ubican cada uno de los
fallecimientos.}

\item{epsg}{Numérico con longitud igual a 1: código EPSG con la proyección de
las coordenadas de la base de datos.}

\item{columnas}{Carácter con longitud igual a 2: nombre de las columnas con
las coordenadas. Debe respetarse su orden, de forma que primero se indique
el nombre de la columna con las longitudes y después el de las latitudes.}

\item{vecinos}{Numérico con longitud igual a 1: número de vecinos más
próximos con los que comparar la mortalidad de cada punto.}

\item{limite}{Numérico con longitud igual a 1: límite de probabilidad.}
}
\description{
Esta función es útil a la hora de comprobar la geocodificación
  de la mortalidad, pues devuelve aquellos puntos con un exceso de
  fallecimientos en relación a la media de sus vecinos más próximos.
}
\details{
La función comienza calculando el número de fallecimientos en cada
  par único de coordenadas (\emph{F}), identifica los \emph{n} vecinos más próximos a cada
  punto y calcula la media de fallecimientos en los mismos (\eqn{\bar{x}}). Considerando que \eqn{F_i \sim Poisson(\bar{x})}, se calcula \eqn{P[F_i < \bar{x}]}.
}
