% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elige_corte}
\alias{elige_corte}
\title{Elige el último punto de corte para los grupos de edad}
\usage{
elige_corte(datos, corte)
}
\arguments{
\item{datos}{Objeto de clase \code{poblaciones_ine}.}

\item{corte}{Numérico: punto de corte (85 o 100).}
}
\value{
Objeto de clase \code{poblaciones_ine} sin la columna
  \code{q_85_plus} en caso de elegir como punto de corte los 100 años de
  edad, o sin las columnas \code{q_85_89:q_100_plus} previo agregado de su
  valor a la columna \code{q_85_plus}.
}
\description{
Elige entre 85 y más (disponible hasta 2010) o 100 y más
  (disponible desde 2011).
}
