% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{elimina_cambios}
\alias{elimina_cambios}
\title{Eliminar cambios manualmente de la base de datos cambios_seccion}
\usage{
elimina_cambios(datos, sc_ref, sc_new)
}
\arguments{
\item{datos}{Objeto de clase \code{cambios_ine} y \code{data.frame} con los
cambios de seccionado.}

\item{sc_ref}{Cadena de caracteres de longitud >= 1 con las secciones de
2011 a eliminar.}

\item{sc_new}{Cadena de caracteres de longitud >= 1 con las secciones del
resto de años a eliminar.}
}
\value{
Un \code{data.frame} sin los cambios seleccionados.
}
\description{
Esta función elimina los cambios especificados del listado de
  cambios de sección detectados.
}
\details{
Los argumentos sc_ref y sc_new deben hacer referencia a una misma
  fila de la base de datos.
}
\seealso{
\code{\link{detecta_cambios}} y \code{\link{une_secciones}}.
}
