% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filtrar_ein_esn}
\alias{filtrar_ein_esn}
\title{Función auxiliar para detectar cambios en secciones censales}
\usage{
filtrar_ein_esn(datos)
}
\arguments{
\item{datos}{objeto de clase \code{tramero_ine}.}
}
\description{
Una función auxiliar para vectorizar el procesado de cambios.
}
