% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\name{geocodificar}
\alias{geocodificar}
\title{Implementación del algoritmo de geocodificación de direcciones}
\usage{
geocodificar(direcciones, codigos_ine = NULL, cartografia = NULL,
  nivel_filtro = c("muni", "prov", "ccaa", "ninguno"))
}
