% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{UTF-8}
\name{geocodificar_google}
\alias{geocodificar_google}
\title{Implementar la cuarta fase del algoritmo de geocodificación de
  direcciones de MEDEA3 (geocodificado con Google)}
\usage{
geocodificar_google(direc, clave_google = NULL, aux.direc = NULL,
  poligono = NULL, varios_resultados = FALSE)
}
\arguments{
\item{direc}{Cadena de carácteres con la dirección a geocodificar.}

\item{clave_google}{Obligatorio: Cadena de carácteres con la clave de uso de
\href{https://cloud.google.com/maps-platform/user-guide/account-changes/}{Google}.}

\item{aux.direc}{Opcional: Lista resultado de la función \code{\link{limpia_dir}}.}

\item{poligono}{Opcional: Objeto de clase \code{\link[sp]{SpatialPolygons}}.}

\item{varios_resultados}{Valor lógico: En caso de encontrar varias coincidencia,
¿desea que se devuelvan todas o solo la primera? Por defecto (FALSE) solo se
devuelve la primera, que es el comportamiento esperado en el protocolo de
geocodificación.}
}
\value{
Un data.frame con una fila y 14 columnas: id, province, muni, tip_via,
  address, portalNumber, refCatastral, postalCode, lat, lng, stateMsg, state,
  type y georef.
}
\description{
Esta función implementa la cuarta fase del algoritmo de
  geocodificación de MEDEA3. En la fase previa se intentó geocodificar las
  direcciones haciendo uso del servicio CartoCiudad
  \code{\link{geocodificar_cartociudad}}. Tras la geocodificación usando
  CartoCiudad, es el momento de probar el motor de Google con las direcciones
  que no hayan sido geocodificadas correctamente.
}
\seealso{
\code{\link{geocodificar_cartociudad}} como paso previo (tercera
  fase del protocolo) y \code{vignette("medear-georreferenciacion")} para
  visualizar el protocolo de georreferenciación.
}
