% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{UTF-8}
\name{limpia_dir}
\alias{limpia_dir}
\title{Limpieza de las cadenas con las direcciones a geocodificar}
\usage{
limpia_dir(tvia, nvia, npoli, muni, prov, codpost)
}
\arguments{
\item{tvia}{Caracter: tipos de vías.}

\item{nvia}{Caracter: nombres de la vías.}

\item{npoli}{Caracter: números de policía.}

\item{muni}{Caracter: nombres de los municipios.}

\item{prov}{Caracter: nombres de las provincias.}

\item{codpost}{Caracter: códigos postales.}
}
\value{
Se devuelve una lista de 6 elementos, uno para cada campo de entrada.
}
\description{
Esta función ejecuta una limpieza genérica de las direcciones,
  eliminando datos perdidos, sustituyendo portales igual a cero o ceros a la
  izquierda, normaliza los pincipales tipos de vía (calle, avenida, plaza,
  partida, camino, carretera, pasaje, paseo y travesía) y elimina
  duplicidades en el nombre de municipios y provincias (consecuencia del uso
  conjunto de varias lenguas).

  La función espera como entrada los campos habituales del boletín de
  defunciones, y ella se ocupa de trabajarlos por separado.
}
\seealso{
\code{\link{geocodificar_cartociudad}} y
  \code{\link{geocodificar_google}}.
}
