% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{latin1}
\name{limpiadirecGoogle}
\alias{limpiadirecGoogle}
\title{Limpieza de caracteres para Google}
\usage{
limpiadirecGoogle(cadena)
}
\arguments{
\item{cadena}{Cadena de caracteres con lsa direcciones a pasar a Google.}
}
\value{
Cadena de caracteres de la misma longitud que la proporcionada.
}
\description{
Eliminación de caracteres no ascii.
}
