% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poblacion}
\alias{poblacion}
\title{Datos de población por sección censal para las ciudades MEDEA3}
\format{Un objeto de clase \code{poblaciones_ine} donde las filas representan
  las distintas secciones censales. Las tres primeras columnas son:
  \describe{
    \item{seccion}{Código de la sección censal en el primer año.}
    \item{sexo}{Código de la sección censal en el segundo año.}
    \item{year}{Primer año.}
  }
  El resto de columnas representan los distintos grupos de edad.}
\usage{
poblacion
}
\description{
Datos de población por sexo y edad (grupos quinquenales) para
  las ciudades implicadas en el proyecto a nivel de sección censal. Los datos
  para algunos años están encriptados por tratarse de consultas específicas
  realizadas al INE y solo son accesibles mediante contraseña. Los datos
  desde 2006 hasta 2017 son de libre acceso siguiendo la licencia del INE, y
  se pueden obtener usando las funciones contenidas en este paquete.
}
\details{
Los códigos de sección censal siguen un orden preestablecido: los
  primeros dos dígitos identifican la provincia, los siguientes tres dígitos
  el municipio, los próximos dos dígitos el distrito y los últimos cuatro a
  la sección censal.

  Hasta el año 2011 el INE agrupa la última categoría de edad como 85 y más,
  mientras que desde el año siguiente llega hasta 100 y más.

  Se ha codificado el sexo como 0 (masculino) y 1 (femenino).
}
\examples{
data(poblacion)
}
\references{
\url{http://www.ine.es/}{Sitio web del INE}.
  \url{http://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177012&menu=resultados&secc=1254736195461&idp=1254734710990}{Población}.
}
\keyword{datasets}
