% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{privacion}
\alias{privacion}
\title{Índices de Privación por Sección Censal para Ciudades MEDEA3}
\format{
Un objeto de clase \code{data.frame}, donde cada
  fila es una sección censal, ofreciendo un valor del índice de privación para 2001 y 2011:
  \describe{
    \item{seccion}{Cádena de 10 caracteres con el código INE de la sección censal.}
    \item{cod_municipio}{Cádena de 5 caracteres con el código INE del municipio.}
    \item{nom_municipio}{Cádena caracteres con el nombre del municipio.}
    \item{privacion_2001}{Valor del índice de privación MEDEA3 en 2001.}
    \item{privacion_2011}{Valor del índice de privación MEDEA3 en 2011.}
  }
}
\usage{
privacion
}
\description{
Índices de privación por sección censal según los censos INE 2001 y 2011,
  calculados siguiendo metodología MEDEA3 (publicación metodológica pendiente) y
   ajustados al seccionado MEDEA3 (tras realizar uniones de secciones 1996-2015).
}
\details{
El índice es comparable entre ciudades y períodos.
}
\examples{

\dontrun{
  library(medear)
  data("privacion")
}
}
\seealso{
\code{\link{carto_medea3}}
}
\keyword{datasets}
