% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{procesa_datos}
\alias{procesa_datos}
\title{Preparar datos para su uso en WinBUGS}
\usage{
procesa_datos(datos)
}
\arguments{
\item{datos}{Lista con (al menos) la cartografía, poblacion y mortalidad, tal
 y como se produce al emplear la función \code{\link{une_secciones}}.
return(list(carto = carto, carto.nb = carto.nb, carto.wb = carto.wb, Obs = Obs, Exp = Exp))}
}
\value{
Devuelve una lista con los siguientes elementos:
  \item{carto}{Cartografía: Objeto de clase \code{\link{une_secciones}} empleando
    proyección WGS84 (EPSG: 4326)}
  \item{carto.nb}{Vecindades: Lista con las vecindades de cada área.}
  \item{carto.wb}{Vecindades WB: Lista con los datos de vecindad necesarios para WinBUGS.}
  \item{Obs}{Matriz tetradimensional (año, sexo, sección censal y causa de defunción)
     de muertes observadas.}
  \item{Exp}{Matriz tetradimensional (año, sexo, sección censal y causa de defunción)
     de muertes esperadas}
}
\description{
Esta función crea los elementos necesarios para ejecutar los modelos
  multivariantes MEDEA3 en WinBUGS.
}
\seealso{
\code{\link{une_secciones}}.
}
