% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{proyecta_secciones}
\alias{proyecta_secciones}
\title{Proyecta pares de coordenadas en un mapa}
\usage{
proyecta_secciones(datos, cartografia, epsg = 4326)
}
\arguments{
\item{datos}{Objeto de clase \code{data.frame} con al menos dos columnas:
lat (latitud) y lng (longitud).}

\item{cartografia}{Objeto de clase \code{\link[sp]{SpatialPolygons}}.}

\item{epsg}{Vector numérico de longitud uno con el código EPSG del sistema de
referencia de coordenadas (CRS) empleado (por defecto se usa el 4326 con
datum WGS84).}
}
\value{
Se devuelve el \code{data.frame} con la columna extra \code{seccion}.
}
\description{
Esta función proyecta las coordenadas (latitud y longitud) en
  un mapa, devolviendo la base de datos origina con una columna extra en la que
  se indica la sección censal en la que recae el punto.
}
\details{
Esta función es capaz de proyectar cualquier base de datos con
  información puntual (mortalidad o población) en la cartografía.
}
\seealso{
\code{\link{une_secciones}}, \code{\link{crea_cubo_mortalidad}} y
  \code{\link{crea_cubo_poblacion}}.
}
