% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{carga_datos}
\alias{carga_datos}
\title{Carga los datos privados de poblacion para el proyecto MEDEA3}
\usage{
carga_datos(key)
}
\arguments{
\item{key}{Cadena de carácteres con la contraseña.}
}
\value{
No se devuelve nada.
}
\description{
Algunos datos del proyecto MEDEA3 están encriptados para poder
  cumplir con la licencia INE (poblaciones desde 1998 a 2003). Esta función
  los desencripta y añade a los datos públicos (resto de años), adjuntando el
  resultado al entorno global.
}
\details{
La contraseña no se almacena en el historial.
}
\examples{

\dontrun{
  carga_datos(key = "contraseña")
}

}
\seealso{
\code{\link{poblacion}} y \code{\link{descarga_poblaciones}}.
}
\keyword{datasets}
