% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{carga_datos}
\alias{carga_datos}
\title{Carga los datos privados del proyecto MEDEA3}
\usage{
carga_datos(key, tipo = c("poblacion", "censo"))
}
\arguments{
\item{key}{Cadena de carácteres con la contraseña.}

\item{tipo}{Tipo de datos a cargar: población o censo.}
}
\value{
Un \code{data.frame} con los datos solicitados.
}
\description{
Algunos datos del proyecto MEDEA3 están encriptados para poder
  cumplir con la licencia INE (poblaciones desde 1998 a 2003 o datos de los
  censos 2001 y 2011). Esta función los desencripta y añade a los datos
  públicos (resto de años).
}
\details{
La contraseña no se almacena en el historial.
}
\examples{

\dontrun{
  # Carga de datos de población
  carga_datos(key = "contraseña", tipo = "poblacion")
}

}
\seealso{
\code{\link{poblacion}} y \code{\link{descarga_poblaciones}}.
}
\keyword{datasets}
