% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cartografia}
\alias{cartografia}
\title{Cartografia por seccion censal para las ciudades MEDEA3}
\format{Un objeto de clase \code{SpatialPoligonsDataFrame}, donde los datos
  (\code{cartografia@data}) tienen clase \code{data.frame} y
  \code{cartografia_ine}, donde cada fila es una sección censal y que cuenta
  con 7 columnas: \describe{ \item{seccion}{Cádena de 10 caracteres con el
  código de sección censal (incluye provincia, municipio, distrito y
  sección).} \item{CUMUN}{Cádena de 5 caracteres con el código INE del
  municipio.} \item{CCA}{Cádena de 2 caracteres con el código INE de la
  comunidad autónoma} \item{NPRO}{Nombre de la provincia.} \item{NCA}{Nombre
  de la comunidad autónoma.} \item{NMUN}{Nombre del municipio.}
  \item{n_viv}{Número de viviendas por sección censal (datos cruzados con
  Dirección General de Catastro).}}}
\usage{
cartografia
}
\description{
Contiene la cartografía por sección censal tal cual puede ser
  utilizada por el paquete \code{sp}.
}
\examples{

\dontrun{
  library(medear)
  library(sp)
  data(cartografia)

  # Representación de los secciones censales de Álava
  plot(cartografia[substring(cartografia$seccion, 1, 5) == "01059", ])

  # Representación de los secciones censales de Álava, según distritos.
  distritos <- substring(cartografia[substring(cartografia$CUSEC, 1, 5) == "01059", ]$CUSEC, 6, 7)
  plot(cartografia[substring(cartografia$CUSEC, 1, 5) == "01059", ], col = as.numeric(distritos))
}
}
\references{
\url{http://www.ine.es/}{ Sitio web del INE}.

  \url{http://www.ine.es/censos2011_datos/cen11_datos_resultados_seccen.htm}{Cartografía}.
}
\keyword{datasets}
