% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{censo}
\alias{censo}
\title{Censos de 2001 y 2011}
\format{Un objeto de clase \code{data.frame} con los siguientes campos:

  \itemize{
    \item seccion Código de sección censal (10 caracteres).
    \item muni Código INE del municipio.
    \item year Año del censo.
    \item i01 Numerador: Trabajadores manuales entre los ocupados, de 16 o más años (dividir por i01_d).
    \item i02 Numerador: Parados de 16 o más años entre los activos (dividir por i02_d).
    \item i03 Numerador: Asalariados eventuales entre los ocupados de 16 o más años (dividir por i02_d).
    \item i04 Numerador: Personas con instrucción insuficiente de 16 o más años (dividir por i03_d).
    \item i05 Numerador: Personas con instrucción insuficiente de 16 a 29 años (dividir por i04_d).
    \item i06 Numerador: Población de 65 o más años (dividir por i05_d).
    \item i07 Numerador: Nacidos en países de renta baja (dividir por i05_d).
    \item i08 Numerador: Nacidos en países de renta baja llegados a España
      después de 2006 (dividir por i05_d).
    \item i01_d Denominador: Población ocupada de 16 o más años.
    \item i02_d Denominador: Población de 16 o más años en situación laboral = 1
      Ocupado/a + 2 Parado que ha trabajado antes + 3 Parado que busca
      empleo (población activa).
    \item i03_d Denominador: Población de 16 o más años.
    \item i04_d Denominador: Población de 16 a 29  años.
    \item i05_d Denominador: Población total.
  }}
\description{
Datos de los censos de 2001 y 2011 necesarios para construir el
  índice de privación. Estos datos se cargan con la función
  \code{\link{carga_datos}}.
}
\examples{

\dontrun{
  library(medear)
  censo <- carga_datos("contraseña", tipo = "censo")
}

}
\seealso{
\code{\link{carga_datos}}
}
\keyword{datasets}
