% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comprueba_geocodificado}
\alias{comprueba_geocodificado}
\title{Comprobacion de asignacion de distintas coordenadas a misma direccion}
\usage{
comprueba_geocodificado(mortalidad, tip_via = FALSE)
}
\arguments{
\item{mortalidad}{Datos con la mortalidad geocodificada. Objeto de clase
\code{data.frame} que contenga, al menos, la siguiente información (los
nombres que deben tener las variables están entre paréntesis): longitud
(\code{lng}), latitud (\code{lat}) y dirección (\code{address}), número de
portal (\code{portalNumber}), municipio (\code{muni}) y provincia
(\code{province}) devueltos por el servicio de geocodificado.}

\item{tipo_via}{Valor Lógico. ¿Debe incorporarse el tipo de vía en la
búsqueda de duplicidades? De ser así la base de datos de mortalidad deberá
incorporar dicha variable (\code{tip_via}).}
}
\value{
Devuelve un \code{data.frame} con la mortalidad y los registros
  problemáticos modificados a \code{NA}, fijando el campo \code{georef} de
  esos registros al valor \code{repetir_geo_google}.
}
\description{
Aunque no es frecuente, en algunas situaciones, y frente a una
  misma dirección, se obtienen distintas coordenadas. Esta función detecta
  estos casos para que pueda volver a lanzarse el algoritmo de geocodificado
  sobre ellos. Concretamente se modifican estos registros de los datos de
  mortalidad, sustituyendo los valores devueltos por el servicio de
  geocodificado por \code{NA}'s, y fijando
  \code{georef == "repetir_geo_google"}.
}
