% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{UTF-8}
\name{comprueba_punto_poligono}
\alias{comprueba_punto_poligono}
\title{Comprobación de inclusión de una coordenada dentro de un polígono}
\usage{
comprueba_punto_poligono(punto, poligono)
}
\arguments{
\item{punto}{Un data.frame con dos columnas: lng y lat.}

\item{poligono}{Un objeto de clase \code{SpatialPolygonsDataFrame} con el
polígono de referencia sobre el que se desee contrastar la pertenecia de
las coordenadas.}
}
\value{
Valor lógico.
}
\description{
Con el fin de asegurar que el resultado de la geocodificación
  sea coherente, esta función permite identificar parejas de coordenadas que
  no estén incluidas en un polígono concreto (una sección, un distrito, un
  municipio o una provincia). Para ello es necesario proporcionar un polígono
  de referencia para cada par de coordenadas.
}
\seealso{
\code{\link{geocodificar_cartociudad}} y \code{\link{geocodificar_google}}
}
