% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{crea_cubo_mortalidad}
\alias{crea_cubo_mortalidad}
\title{Crear la matriz 5-dimensional de mortalidad}
\usage{
crea_cubo_mortalidad(datos, cartografia, epsg = 4326, medea3 = TRUE,
  otras_causas = NULL, corte_edad = 85, periodo = 1996:2015)
}
\arguments{
\item{datos}{Objeto de clase \code{data.frame} con la mortalidad
geocodificada. Este objeto debe contener las variables \code{sexo},
\code{year_defuncion}, \code{edad}, \code{causa_defuncion} \code{lng} y
\code{lat}, con exactamente esos nombres.}

\item{cartografia}{Objeto de clase \code{\link[sp]{SpatialPolygons}} con
proyección asignada (código EPSG).}

\item{epsg}{Vector numérico de longitud uno con el código EPSG del sistema de
referencia de coordenadas (CRS) empleado en los datos de mortalidad (por
defecto se usa el 4326 con datum WGS84).}

\item{medea3}{¿Se desea utilizar las 22 agrupaciones MEDEA3? Por defecto sí.}

\item{otras_causas}{Véctor de caracteres que indica los nombres de las
columnas (columnas con valor 0-1) en la base de datos de mortalidad que
identifican a dichas otras causas.}

\item{corte_edad}{Numérico: punto de corte para los grupos de edad (85 o
100). Argumento opcional en caso de proporcionar datos de poblaciones o
mortalidad.}

\item{periodo}{Vector numérico de longitud >= 1 con los años para los que se
desee construir el \code{array} de mortalidad.}
}
\description{
Esta función crea la matriz 5-dimensional de mortalidad (año de
  defunción, sexo, grupo de edad quinquenal, sección censal y causa de
  defunción).
}
\seealso{
\code{\link{une_secciones}} y \code{\link{proyecta_secciones}}.
}
