% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{crea_cubo_poblacion}
\alias{crea_cubo_poblacion}
\title{Crear la matriz 4-dimensional de poblaciones}
\usage{
crea_cubo_poblacion(datos, corte_edad = 85, periodo = 1996:2015,
  datos_propios = FALSE)
}
\arguments{
\item{datos}{Objeto de clase \code{data.frame} con las cifras de población.
Este objeto debe contener las variables \code{seccion},
\code{sexo}, \code{year} y los grupos quinquenales desde \code{q_0_4}
hasta \code{q_85_plus} o \code{q_100_plus}, con exactamente esos nombres.}

\item{corte_edad}{Numérico: punto de corte para los grupos de edad (85 o
100). Argumento opcional en caso de proporcionar datos de poblaciones o
mortalidad.}

\item{periodo}{Vector numérico de longitud >= 1 con los años para los que se
desee construir el \code{array} de mortalidad.}

\item{datos_propios}{Valor lógico. Indica si se dispone de datos de población
a nivel puntual. Por defecto tiene tiene el valor \code{FALSE}. En caso
contrario la función devolvería el \code{array} de poblaciones pero con
\code{NA} en todos los conteos, con lo que el usuario será el encargado de
rellenar este \code{array} posteriormente.}
}
\description{
Esta función crea la matriz 4-dimensional con las poblaciones
  (año, sexo, grupo de edad quinquenal y sección censal).
}
\seealso{
\code{\link{une_secciones}} y \code{\link{crea_cubo_mortalidad}}.
}
