% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\encoding{UTF-8}
\name{descarga_cartografia}
\alias{descarga_cartografia}
\title{Funcion para descargar la cartografia con el seccionado del INE para
  2011}
\usage{
descarga_cartografia(crs = 4326, conservar = TRUE)
}
\arguments{
\item{crs}{Vector numérico de longitud uno con el código EPSG del sistema de
referencia de coordenadas (CRS) empleado (por defecto se usa el 4326 con
datum WGS84).}

\item{conservar}{Valor lógico: ¿se desea conservar los archivos descargados
en el directorio oculto \code{./.cartografia/} dentro del directorio de
trabajo?}
}
\value{
Un objeto de clase \code{SpatialPolygonsDataFrame}, donde cada fila
  es una sección censal y que cuenta con 7 columnas: \item{seccion}{Cadena de
  10 caracteres con el código de sección censal (incluye provincia, municipio
  y distrito).} \item{CUMUN}{Cadena de 5 caracteres con el código del
  municipio (incluye provincia).} \item{CCA}{Cadena de 2 caracteres con el
  código de comunidad autónoma.} \item{NPRO}{Nombre de la provincia.}
  \item{NCA}{Nombre de la comunidad autónoma.} \item{NMUN}{Nombre del
  municipio.} \item{geometry}{Columna de tipo lista con la geometría asociada
  a cada sección censal.}
}
\description{
Descarga la cartografía ofrecida públicamente por el INE para el
  año 2011.
}
\details{
Aunque el INE emplea otro CRS, se recomienda utlizar el CRS 4326
  como elemento normalizado.
}
\examples{

\dontrun{
  library(medear)
  library(sp)
  carto_ine    <- descarga_cartografia()
  carto_ine_46 <- carto[substr(carto$seccion, 3, 4) == "46", ]
  plot(carto_ine_valencia)
}

}
\seealso{
\code{\link{descarga_trameros}} y \code{\link{descarga_poblaciones}}
}
