% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\encoding{UTF-8}
\name{descarga_cartografia}
\alias{descarga_cartografia}
\title{Función para descargar la cartografía con el seccionado del INE para
  2011}
\usage{
descarga_cartografia(crs = 4326, conservar = TRUE)
}
\arguments{
\item{crs}{Vector numérico de longitud uno con el código EPSG del sistema de
referencia de coordenadas (CRS) empleado (por defecto se usa el 4326 con
datum WGS84).}

\item{conservar}{Valor lógico: ¿se desea conservar los archivos descargados
en el directorio oculto \code{cartografia/} dentro del directorio de
trabajo?}
}
\value{
Un objeto de clase \code{cartografia_ine} y \code{sf}, donde cada
  fila es una sección censal y que cuenta con 13 columnas:
  \item{CUSEC}{Cádena de 10 caracteres con el código de sección censal
  (incluye provincia, municipio y distrito).}
  \item{CUMUN}{Cádena de 5 caracteres con el código del municipio (incluye
  provincia).}
  \item{CSEC}{Cádena de 3 caracteres con el código de sección censal.}
  \item{CDIS}{Cádena de 2 caracteres con el código de distrito.}
  \item{CPRO}{Cádena de 3 caracteres con el código de provincia.}
  \item{CCA}{Cádena de 2 caracteres con el código de comunidad autónoma.}
  \item{CUDIS}{Cádena de 7 caracteres con el código de distrito (incluye
  provincia y  municipio).}
  \item{OBS}{Observaciones por parte del proveedor de los datos.}
  \item{NPRO}{Nombre de la provincia.}
  \item{NCA}{Nombre de la comunidad autónoma.}
  \item{NMUN}{Nombre del municipio.}
  \item{geometry}{Columna de tipo lista con la geometría asociada a cada
  sección censal.}
}
\description{
Descarga la cartografía ofrecida públicamente por el INE para el
  año 2011.
}
\details{
Aunque el INE emplea otro CRS, se recomienda utlizar el CRS 4326.
}
\examples{
\dontrun{
library(sf)
carto_ine    <- descarga_cartografia()
carto_ine_46 <- carto[, "CPRO" == "46"]
plot(st_geometry(carto_ine_valencia))
}

}
