% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\encoding{UTF-8}
\name{descarga_poblaciones}
\alias{descarga_poblaciones}
\title{Descarga poblaciones del INE por sección censal, sexo, edad y año}
\usage{
descarga_poblaciones(cod_provincia = c(paste0("0", 1:9), 10:52), years
  = 2006:2015, descarga = TRUE, ruta = NULL, conservar = TRUE, ntries = 10)
}
\arguments{
\item{cod_provincia}{Cadena de carácteres de longitud >= 1 con el código de
la/s provincia/s en las que se desee obtener el listado de cambios de
seccionado.}

\item{years}{Vector numérico de longitud >= 1 con los años para los que se
desee consultar las variaciones de seccionado.}

\item{descarga}{Valor lógico: ¿debe procederse a la descarga de los datos?}

\item{ruta}{Cadena de carácteres indicando la ruta en la que se almacenan los
archivos tal cual se descargaron desde el INE, en caso de escoger
\code{descarga = FALSE}.}

\item{conservar}{Valor lógico: ¿se desea conservar los archivos descargados
en el directorio oculto \code{./.poblaciones/} dentro del directorio de
trabajo?}

\item{ntries}{Valor numérico: número de intentos en caso de mala conexión.}
}
\value{
Un objeto de clase \code{poblaciones_ine} donde las filas representan
  las distintas secciones censales. Las tres primeras columnas son:
  \item{seccion}{Código de la sección censal.} \item{sexo}{Codificado como 0
  para hombres y 1 para mujeres.} \item{year}{Año al que se hace referencia.}
  El resto de columnas representan los distintos grupos de edad.
}
\description{
Descarga o carga las poblaciones anuales del INE por sección
  censal, sexo y edad por grupos quinquenales (datos desde 2006 en adelante).
}
\details{
El tiempo de ejecución de la función varía según el número de
  provincias y el rango de años. La forma más sencilla de acelerar el proceso
  de computación es mediante la ejecución en paralelo de la función.

  Los códigos de sección censal siguen un orden preestablecido: los primeros
  dos dígitos identifican la provincia, los siguientes tres dígitos el
  municipio, los próximos dos dígitos el distrito y los últimos tres a la
  sección censal.

  Hasta el año 2011 el INE agrupa la última categoría de edad como 85 y más,
  mientras que desde el año siguiente llega hasta 100 y más.

  Si se desea acceder a las poblaciones desde 1996 (datos adquiridos al INE),
  se debe utilizar la función \code{\link{carga_datos}}.
}
\examples{

\dontrun{
  library(medear)
  poblaciones <- descarga_poblaciones(cod_provincia = "46")
  poblaciones
}

}
\seealso{
\code{\link{carga_datos}}, \code{\link{descarga_trameros}} y
  \code{\link{descarga_cartografia}}.
}
