% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\encoding{UTF-8}
\name{descarga_trameros}
\alias{descarga_trameros}
\title{Descarga los trameros del INE}
\usage{
descarga_trameros(cod_provincia = c(paste0("0", 1:9), 10:52), years =
  c(2001, 2004:2016), descarga = TRUE, ruta = NULL, conservar = TRUE)
}
\arguments{
\item{cod_provincia}{Cadena de carácteres de longitud >= 1 con el código de
la/s provincia/s en las que se desee obtener el listado de cambios de
seccionado.}

\item{years}{Vector numérico de longitud >= 2 con los años para los que se
desee consultar las variaciones de seccionado.}

\item{descarga}{Valor lógico: ¿debe procederse a la descarga de los trameros?}

\item{ruta}{Cadena de carácteres indicando la ruta en la que se almacenan los
archivos tal cual se descargaron desde el INE, en caso de escoger
\code{descarga = FALSE}.}

\item{conservar}{Valor lógico: ¿se desea conservar los archivos descargados
en el directorio oculto \code{./.trameros/} dentro del directorio de
trabajo?}

\item{...}{Parámetros adicionales en la lectura de datos (no es necesario su
uso).}
}
\value{
Un objeto de clase \code{tramero_ine} con 11 columnas:
  \item{CPRO}{Código de la provincia.} \item{CMUM}{Código del municipio.}
  \item{DIST}{Código del distrito.} \item{SECC}{Código de la sección censal
  reducido.} \item{CVIA}{Código de la vía reducido.} \item{EIN}{Primer portal
  del tramo de vía.} \item{ESN}{Último portal del tramo de vía.}
  \item{NVIAC}{Nombre de la vía.} \item{seccion}{Código de la sección censal
  completo.} \item{year}{Año del tramero.} \item{via}{Código de la vía
  completo.}

  Cada fila representa un tramo de vía, puediendo repetirse la misma vía en
  varias ocasiones en función de si su recorrido recae en varias secciones
  censales, o si se trata de tramos de la vía con numeración par, impar o sin
  numeración alguna.
}
\description{
Descarga los trameros que ofrece al público el INE para el año
  2001 y desde 2004 a 2016.
}
\details{
El tiempo de ejecución de la función varía según el número de
  provincias y el rango de años. La forma más sencilla de acelerar el proceso
  de computación es mediante la ejecución en paralelo de la función.

  Los códigos de sección censal siguen un orden preestablecido: los primero
  dos dígitos identifican la provincia, los siguientes tres dígitos el
  municipio, los próximos dos dígitos el distrito y los últimos tres dígitos
  la sección censal.
}
\examples{
\dontrun{
  library(medear)
  trameros <- descarga_trameros(cod_provincia = c("51", "52"))
  trameros
}

}
\seealso{
\code{\link{descarga_cartografia}} y
  \code{\link{descarga_poblaciones}}.
}
