% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{detecta_cambios}
\alias{detecta_cambios}
\title{Funcion para detectar cambios de seccionado en trameros}
\usage{
detecta_cambios(datos, years = c(1996, 2001, 2004:2016))
}
\arguments{
\item{datos}{Objeto de clase \code{tramero_ine} (devuelto por la función
\code{\link{descarga_trameros}}), incluyendo obligatoriamente al año 2011.}

\item{years}{Vector numérico de longitud >= 2 con los años para los que se
desee consultar las variaciones de seccionado.}
}
\value{
Un objeto de clase \code{cambios_ine} con 5 columnas:
  \item{sc_ref}{Código de la sección censal en el primer año.}
  \item{sc_new}{Código de la sección censal en el segundo año.}
  \item{year}{Primer año.} \item{year2}{Segundo año.} \item{vias}{Lista con
  el código de las vías que provocan el cambio de sección (incorporando un
  dígito al final de la cadena indicando si se trata de numeración par (0) o
  impar(1))}
}
\description{
Detecta cambios de sección censal para las provincias y el
  período marcados.
}
\details{
El tiempo de ejecución de la función varía según el número de
  provincias y el rango de años. La forma más sencilla de acelerar el proceso
  de computación es mediante la ejecución en paralelo de la función.

  Los códigos de sección censal siguen un orden preestablecido: los primeros
  dos dígitos identifican la provincia, los siguientes tres dígitos el
  municipio, los próximos dos dígitos el distrito y los últimos tres dígitos
  hacen referencia a la sección censal.
}
\examples{

\dontrun{
  library(medear)
  trameros <- descarga_trameros(cod_provincia = c("51", "52"))
  cambios  <- detecta_cambios(datos = trameros)
  cambios
}

}
\seealso{
\code{\link{une_secciones}} y \code{\link{descarga_trameros}}.
}
