% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{detecta_cluster}
\alias{detecta_cluster}
\title{Deteccion de agrupaciones de mortalidad a revisar manualmente}
\usage{
detecta_cluster(datos, epsg = 4326, vecinos = 10, cartografia = NULL,
  limite = c(1e-10, 1e-15, 1e-20))
}
\arguments{
\item{datos}{Base de datos con las coordenadas que ubican cada uno de los
fallecimientos. Debe contener, al menos, 3 columnas: \code{BOD.direccion},
\code{lat} y \code{lng}, las cuales deben tener exactamente esos nombres.
De forma paralela la función tratará de buscar otras seis columnas fruto
del protocolo de geocodificación (\code{province}, \code{muni},
\code{tip_via}, \code{address}, \code{portalNumber} y \code{postalCode})
(nuevamente deben tener exactamente esos nombres, que son los que resultan
del protocolo de geocodificación), con el propósito de mostrar, junto a las
direcciones del BOD, las direcciones devueltas por los servicios de
geocodificado. Si la base de datos tuviera otros nombres, seria trabajo del
usuario cambiárselos como paso previo al uso de la función.}

\item{epsg}{Numérico con longitud igual a 1: código EPSG con la proyección de
las coordenadas de la base de datos. Por defecto se usa el EPSG 4326
(longlat WGS 84).}

\item{vecinos}{Numérico con longitud igual a 1: número de vecinos más
próximos con los que comparar la mortalidad de cada punto. Por defecto 10.}

\item{cartografia}{Objeto de clase \code{\link[sp]{SpatialPolygons}}. Con
ella se representa el seccionado oficial de 2011 a modo de referencia con
la que juzgar la geocodificación de un conjunto de defunciones anómalas. En
principio tiene valor nulo, lo que implica usar la cartografía asociada al
paquete para las ciudades MEDEA3. Si el usuario desea consultar otros
municipios puede hacer uso de este argumento para tener el seccionado de
fondo.}

\item{limite}{Numérico con longitud 1 <= x <= 9: límites de probabilidad con
los que se identifican las agrupaciones sospechosas. Por defecto viene
fijado a 1e-10, 1e-15 y 1e-20.}
}
\value{
Representa un objeto de clase \code{\link[leaflet]{leaflet}} en el
  que se marcan los puntos a revisar. Por defecto los puntos se dividen en
  tres colores en función de la mortalidad acontecida en las coordenadas
  vecinas: verde si la probabilidad de obtener ese resultado es inferior a
  1e-10, azul si es inferior a 1e-15 y rojo si es inferior a 1e-20 (el caso
  más evidente). Haciendo clic en cada uno de los puntos se puede consultar
  las direcciones, tanto del BOD como las obtenidas en la geocodificación (se
  puede cambiar entre una y otra en el menú del extremo superior derecho),
  asociadas a cada punto. Al comienzo de cada dirección se indica el número
  de fila al que hace referencia cada dirección en los datos que se
  facilitaron, de forma que pueda recuperarse fácilmente dicha información
  para explorar los datos en profundidad.

  Respecto al menú anteriormente mencionado, también permite cambiar la capa
  de visualización de fondo, utilizando OpenStreetMap, Google Maps o Google
  Satellite (de forma que pueda disponerse de información extra sin salir de
  la aplicación).

  Si se desea efectuar el cálculo de distancias o áreas, en la esquina
  inferior izquierda se dispone de un menú para realizarlas (en metros y en
  metros cuadrados).

  La función también devuelve los datos con las cooredenadas (campos
  \code{lng} y \code{lat}), número de defunciones (campo \code{N}), límite
  dentro del cual recae (campo \code{limite}, útil para filtrar resultados) y
  direcciones asociadas a cada punto (campo \code{detalle}, de tipo lista
  donde cada elemento contiene un data.frame con la información detallada).
}
\description{
Esta función es útil a la hora de comprobar la geocodificación
  de la mortalidad, pues devuelve aquellos puntos con un exceso de
  fallecimientos en relación a la media de sus vecinos más próximos. Con esta
  dinámica es fácil identificar los centros residenciales o errores
  sistemáticos en la geocodificación, donde es de esperar una mayor
  aglomeración de defunciones.
}
\details{
La función comienza calculando el número de fallecimientos en cada
  par único de coordenadas, identifica los \emph{n} vecinos más próximos a
  cada punto y calcula la media de fallecimientos en los mismos. Considerando
  que los fallecimientos en cada par de coordenadas siguen una distribución
  de Poisson cuya media es la media de los fallecimientos en los puntos más
  cercanos, se calcula la probabilidad de superar esa media.
}
\examples{
\dontrun{
  library(medear)
  revisar <- detecta_cluster(datosmort)
  revisar
}

}
