% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{UTF-8}
\name{filtra_dir}
\alias{filtra_dir}
\title{Deteccion y prueba de variantes de direcciones mal escritas}
\usage{
filtra_dir(vias, nivel)
}
\arguments{
\item{vias}{Lista de seis elementos devuelta por \code{\link{limpia_dir}}.}

\item{nivel}{Numérico: filtro a a aplicar.}
}
\value{
Se devuelve un vector de carácteres de igual longitud a la lista de
  entrada, con las direcciones listas para geocodificar (las direcciones en
  las que no se reconozca ningún patrón son sustituidas por un elemento
  vacío).
}
\description{
Se considera una serie de variantes de las cadenas de carácteres
  de las direcciones no geocodificadas. La intención es valorar si esas
  variaciones podrían producir en alguna ocasión una geocodificación exitosa.
  En concreto, \code{\link{filtra_dir}} contempla 5 posibles variantes para
  las direcciones que no han podido ser geocodificadas:
  \enumerate{\item{eliminar duplicidad de tipos de vía (ejemplo: calle camino
  ...-> camino ...);} \item{eliminar descripciones de vía (ejemplo: Avenida
  rosa (Edificio azul)->Avenida rosa);} \item{eliminar palabras de 3 o menos
  carácteres (ejemplo: calle la marina alta-> calle marina alta);}
  \item{eliminar signos de puntuación (ejemplo: calle gral. pedro.->calle
  gral pedro);} \item{implementación de todas las variantes anteriores de
  forma secuencial.}} \code{\link{filtra_dir}} contempla todas estas
  variantes para cualquier dirección que no haya podido ser geocodificada a
  partir de su dirección original.
}
\seealso{
\code{\link{limpia_dir}}, \code{\link{geocodificar_cartociudad}} y
  \code{\link{geocodificar_google}}.
}
