% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{UTF-8}
\name{geocodificar_cartociudad}
\alias{geocodificar_cartociudad}
\title{Implementar la tercera fase del algoritmo de geocodificación de
  direcciones de MEDEA3 (geocodificado con CartoCiudad)}
\usage{
geocodificar_cartociudad(direc, poligono = NULL)
}
\arguments{
\item{direc}{Cadena de carácteres con laa direcciones a georreferenciar.}

\item{poligono}{Opcional: objeto de clase
\code{\link[sp]{SpatialPolygons}}.}
}
\value{
Un data.frame con tantas filas como direcciones se haya proporcionado
  y 14 columnas: id, province, muni, tip_via, address, portalNumber,
  refCatastral, postalCode, lat, lng, stateMsg, state, type y georef.
}
\description{
Esta función implementa la tercera parte del algoritmo de
  geocodificación de MEDEA3. En primer lugar se intenta geocodificar las
  direcciones haciendo uso del servicio CartoCiudad en su
  \href{http://www.cartociudad.es/CartoGeocoder/Geocode}{versión antigua}. En
  esta primera parte daremos por válidos todas aquellos direcciones que hayan
  obtenido estado == 1 (se ha encontrado la dirección correspondiente de
  forma exacta) o estado == 2 (dirección asignada al portal más próximo).
  Tras esto se intentará geocodificar con
  \href{http://www.cartociudad.es/geocoder/api/geocoder/findJsonp}{la nueva
  versión de CartoCiudad} al resto de direcciones no geocodifadas por la
  versión anterior, junto con las direcciones geocodificadas que no tuvieran
  portal y las que hayan obtenido status == 2. La geocodificación de las
  direcciones que no tienen portal resulta menos fiable en la versión antigua
  de CartoCiudad ya que son situadas en el inicio de su vía, mientras que en
  la nueva versión se sitúan en el centro, siendo este abordaje más acertado.
  Por otro lado, la versión antigua de CartoCiudad en ocasiones cambia de
  acera (numeros pares a impares y viceversa) algunas direcciones mientras
  que en la versión nueva esto no ocurre. Por ello, intentaremos solucionar
  estas dos situaciones con la nueva versión de CartoCiudad, y en caso de que
  esta no logre obtener una ubicación satisfactoria mantendríamos la
  geocodificación original conseguida por la versión previa.

  Tras el proceso descrito, pueden considerarse una serie de variantes de las
  cadenas de carácteres de las direcciones no geocodificadas, empleando la
  función \code{\link{filtra_dir}}.

  La función incorpora un filtro cartográfico que deseche aquellas
  localizaciones que caigan fuera de un polígono concreto, para lo cual es
  necesario incluirlo como argumento de la función.

  Tras la geocodificación usando CartoCiudad, es el momento de probar el
  motor de Google con las direcciones que no hayan sido geocodificadas
  correctamente.
}
\seealso{
\code{\link{geocodificar_google}} para georreferenciar los registros
  pendientes (cuarta fase del protocolo) y
  \code{vignette("medear-georreferenciacion")} para visualizar el protocolo
  de georreferenciación.
}
