% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodificar.R
\encoding{UTF-8}
\name{limpiadirecGoogle}
\alias{limpiadirecGoogle}
\title{Limpieza de carácteres para Google}
\usage{
limpiadirecGoogle(cadena)
}
\arguments{
\item{cadena}{Cadena de carácteres con las direcciones a pasar a Google.}
}
\value{
Cadena de carácteres de la misma longitud que la proporcionada.
}
\description{
Eliminación de carácteres no ASCII para la geocodificación con
  Google.
}
\seealso{
\code{\link{geocodificar_google}}.
}
