% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secciones}
\alias{secciones}
\title{Secciones censales únicas por año (1996-2016)}
\format{Un objeto de clase \code{data.frame} y \code{data.table}, donde cada
  fila es una sección para un año determinado.}
\usage{
secciones
}
\description{
Listado con las secciones censales únicas por año (1996-2016).
}
\examples{

\dontrun{
  library(medear)
  data("secciones")
}
}
\keyword{datasets}
