% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_cambios.R
\encoding{UTF-8}
\name{une_secciones}
\alias{une_secciones}
\title{Une los cambios del seccionado del INE}
\usage{
une_secciones(cambios, cartografia, years = 1996:2016, poblacion =
  NULL, corte_edad = 85, catastro = FALSE, umbral_vivienda = 5, umbral_tramo
  = 10)
}
\arguments{
\item{cambios}{Objeto de clase \code{cambios_ine}.}

\item{cartografia}{Objeto de clase \code{\link[sp]{SpatialPolygons}}, y con
datos de clase \code{cartografia_ine}.}

\item{years}{Vector numérico de longitud >= 1 con los años para los que se
desee consultar las variaciones de seccionado. El año 2011 debe figurar
dentro del vector, cuyo rango debe ser continuo (sin saltos de más de un
año).}

\item{poblacion}{Objeto de clase \code{poblaciones_ine}. Argumento opcional a
proporcionar en caso de querer agregar las poblaciones.}

\item{corte_edad}{Numérico: punto de corte para los grupos de edad (85 o
100). Argumento opcional en caso de proporcionar datos de poblaciones.}

\item{catastro}{Lógico: ¿El archivo de cambios incorpora datos sobre el
catastro? Por defecto \code{catastro = FALSE}.}

\item{umbral_vivienda}{Numérico: porcentaje de viviendas afectadas en el
cambio de sección. Solo se utiliza si \code{catastro = TRUE}. Por defecto
se fija al 5 \%.}

\item{umbral_tramo}{Numérico: porcentaje de tramos afectados por la unión de
secciones respecto al total de tramos contenidos en la sección de
referencia (2011). Solo se utiliza si \code{catastro = TRUE}. Por defecto
se fija a 10 \%.}
}
\value{
El resultado devuelto varía en función de si se proporcionan datos de
  poblaciones o no. Si no se proporcionan se devuelve un objeto de clase
  \code{cartografia_ine} y \code{\link[sp]{SpatialPolygonsDataFrame}} con la
  cartografía, donde cada fila es una sección censal y que cuenta con 9
  columnas: \describe{\item{seccion}{Cadena de 10 caracteres con el código de
  sección censal (incluye provincia, municipio y distrito).}
  \item{CUMUN}{Cadena de 5 caracteres con el código del municipio (incluye
  provincia).} \item{CCA}{Cadena de 2 caracteres con el código de comunidad
  autónoma.} \item{NPRO}{Nombre de la provincia.} \item{NCA}{Nombre de la
  comunidad autónoma.} \item{NMUN}{Nombre del municipio.}
  \item{geometry}{Columna de tipo lista con la geometría asociada a cada
  sección censal.} \item{cluster_id}{Código de identificación del cluster de
  uniones.} \item{sc_unida}{Código de las secciones unidas.}}

  En caso de proporcionan poblaciones, se devuelve una lista de longitud
  igual a dos, donde el primer elemento es la cartografía descrita
  anteriormente y el segundo elemento de la lista es un objeto de clase
  \code{poblaciones_ine} donde las filas representan las distintas secciones
  censales. Las tres primeras columnas son: \describe{\item{seccion}{Código
  de la sección censal.} \item{sexo}{Sexo de la población (0 = masculino; 1 =
  femenino).} \item{year}{Año de referencia.}} El resto de columnas
  representan los distintos grupos de edad, tras realizar el corte en los
  grupos de edad (85 0 100).
}
\description{
Une los cambios del seccionado del INE (tomando como referencia
  la cartografía INE 2011), adaptando a su vez las poblaciones por sexo año y
  sección censal. Si el archivo de cambios incorpora información catastral
  (número de viviendas afectada por cada cambio de sección), se puede fijar
  un umbral de cambio (\%) para rechazar aquellos cambios que involucren a
  muy pocas viviendas.
}
\details{
Cuando se quiera unir cambios no solo en la cartografía sino también
  en los datos de población, puede aparecer un comportamiento inestable de la
  función. Esto es debido divergencias existentes en la información contenida
  en los datos de población y en los trameros (que es desde donde se crea el
  listado de cambios de sección), a pesar de que en ambos casos la fuente de
  información es el propio INE.

  Lo anterior se traduce en que, para determinadas consultas, el número de
  secciones contenidas en los datos de cartografía y poblaciones no será el
  mismo, siendo lo más habitual que esto ocurra con los datos de población.
  Cuando esto pase (si pasa) la función devolverá un aviso, indicando qué
  secciones se ven afectadas y en qué años, de forma que el usuario pueda
  tratar de solucionarlo por su cuenta, aunque no hay una solución perfecta.

  Las dos soluciones más efectivas (aunque son soluciones \emph{ad hoc} y
  recae en el usuario encontrar la más apropiada para su consulta concreta)
  que se han encontrado son:

  \enumerate{ \item modificar los criterios temporales de la consulta,
  ampliando o reduciendo el marco temporal (p. ej., pasar de un período
  2001:2015 a 1996:2015 o 2002:2014); \item consultar las secciones
  problemáticas (accesibles mediante la consulta \code{attr(yy,
  "sc_not_in_cartografia")}) en los datos de población y, en base al archivo
  de cambios de sección, decidir con qué sección se debería realizar la
  unión.}

  En el ejemplo se desarrollan ambos abordajes, con un tratamiento más
  extensivo en la viñeta de unión de seccionado (XXXXXXXXX).
}
\examples{

\dontrun{
  # En este ejemplo se trabaja con la ciudad de Palma de Mallorca (código ine: 07040)
  # en los años 2004-2015

  library(medear)
  data("poblacion")
  data("cambios_seccion")
  data("cartografia")
  cambios_pm     <- cambios_seccion[substr(cambios_seccion$sc_ref, 1, 5) == "07040", ]
  cartografia_pm <- cartografia[cartografia$CUMUN == "07040", ]
  poblacion_pm   <- poblacion[substr(poblacion$seccion, 1, 5) == "07040", ]

  ##########################################################################
  ## Ejemplo sin utilizar el filtro de catastro                           ##
  ##########################################################################

  union_sin_cat <- une_secciones(
    cambios         = cambios_pm,
    cartografia     = cartografia_pm,
    years           = 2004:2015,
    poblacion       = poblacion_pm,
    catastro        = FALSE
  )

  nrow(union_sin_cat$cartografia) # 197 secciones
  length(unique(union_sin_cat$poblacion$seccion)) # 197 secciones
  round(nrow(union_sin_cat$cartografia) / nrow(cartografia_ali), 2) # Conserva el 78 \\\% de secciones
  # No salta ningún aviso sobre divergencia en seccionado entre cartografía y poblaciones,
  # pero se puede comprobar manualmente:
  unique(union_sin_cat$poblacion$seccion)[!unique(union_sin_cat$poblacion$seccion) \%in\% union_sin_cat$cartografia$seccion]
  union_sin_cat$cartografia$seccion[!union_sin_cat$cartografia$seccion \%in\% unique(union_sin_cat$poblacion$seccion)]
  # No hay ninguna sección problemática.


  ##########################################################################
  ## Ejemplo utilizando el filtro de catastro                             ##
  ##########################################################################

  union_con_cat <- une_secciones(
    cambios         = cambios_pm,
    cartografia     = cartografia_pm,
    years           = 2004:2016,
    poblacion       = poblacion_pm,
    catastro        = TRUE,
    umbral_vivienda = 5,
    umbral_tramo    = 10
  )

  nrow(union_con_cat$cartografia) # 197 secciones
  length(unique(union_con_cat$poblacion$seccion)) # 197 secciones
  round(nrow(union_con_cat$cartografia) / nrow(cartografia_ali), 2) # Conserva el 78 \\\% de secciones
  # No salta ningún aviso sobre divergencia en seccionado entre cartografía y poblaciones,
  # pero se puede comprobar manualmente:
  unique(union_con_cat$poblacion$seccion)[!unique(union_con_cat$poblacion$seccion) \%in\% union_con_cat$cartografia$seccion]
  union_con_cat$cartografia$seccion[!union_con_cat$cartografia$seccion \%in\% unique(union_con_cat$poblacion$seccion)]
  # No hay ninguna sección problemática.





  poblacion   <- uniones$poblacion
  cartografia <- uniones$cartografia
}

}
\seealso{
\code{\link{detecta_cambios}}, \code{\link{descarga_poblaciones}} y
  \code{\link{descarga_cartografia}}
}
