% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_cambios.R
\encoding{UTF-8}
\name{une_secciones}
\alias{une_secciones}
\title{Une los cambios del seccionado del INE}
\usage{
une_secciones(cambios, cartografia, years = 1996:2016, poblacion =
  NULL, corte_edad = 85, catastro = FALSE, umbral_cambio = 10, distancia_max
  = 100)
}
\arguments{
\item{cambios}{Objeto de clase \code{cambios_ine}.}

\item{cartografia}{Objeto de clase
\code{\link[sp]{SpatialPolygonsDataFrame}}, y con datos de clase
\code{cartografia_ine}.}

\item{years}{Vector numérico de longitud >= 1 con los años para los que se
desee consultar las variaciones de seccionado. El año 2011 debe figurar
dentro del vector, cuyo rango debe ser continuo (sin saltos de más de un
año).}

\item{poblacion}{Objeto de clase \code{poblaciones_ine}. Argumento opcional a
proporcionar en caso de querer agregar las poblaciones.}

\item{corte_edad}{Numérico: punto de corte para los grupos de edad (85 o
100). Argumento opcional en caso de proporcionar datos de poblaciones.}

\item{catastro}{Lógico: ¿El archivo de cambios incorpora datos sobre el
catastro? Por defecto \code{catastro = FALSE}.}

\item{umbral_cambio}{Numérico: porcentaje de viviendas afectadas en el cambio
de sección. Solo se utiliza si \code{catastro = TRUE}. Por defecto se fija
al 10 %.}

\item{distancia_max}{Numérico: máxima distancia (en metros) a la que pueden
estar dos secciones para ser unidas. Por defecto se fija a 100 m.}
}
\value{
El resultado devuelto varía en función de si se proporcionan datos de
  poblaciones o no. Si no se proporcionan se devuelve un objeto de clase
  \code{cartografia_ine} y \code{\link[sp]{SpatialPolygonsDataFrame}} con la
  cartografía, donde cada fila es una sección censal y que cuenta con 9
  columnas: \item{seccion}{Cadena de 10 caracteres con el código de sección
  censal (incluye provincia, municipio y distrito).} \item{CUMUN}{Cadena de 5
  caracteres con el código del municipio (incluye provincia).}
  \item{CCA}{Cadena de 2 caracteres con el código de comunidad autónoma.}
  \item{NPRO}{Nombre de la provincia.} \item{NCA}{Nombre de la comunidad
  autónoma.} \item{NMUN}{Nombre del municipio.} \item{geometry}{Columna de
  tipo lista con la geometría asociada a cada sección censal.}
  \item{cluster_id}{Código de identificación del cluster de uniones.}
  \item{sc_unida}{Código de las secciones unidas.}

  En caso de proporcionan poblaciones, se devuelve una lista de longitud
  igual a dos, donde el primer elemento es la cartografía descrita
  anteriormente y el segundo elemento de la lista es un objeto de clase
  \code{poblaciones_ine} donde las filas representan las distintas secciones
  censales. Las tres primeras columnas son: \item{seccion}{Código de la
  sección censal en el primer año.} \item{sexo}{Código de la sección censal
  en el segundo año.} \item{year}{Primer año.} El resto de columnas
  representan los distintos grupos de edad, tras realizar el corte en los
  grupos de edad (85 0 100).
}
\description{
Une los cambios del seccionado del INE (tomando como referencia
  la cartografía INE 2011), adaptando a su vez las poblaciones por sexo año y
  sección censal. Si el archivo de cambios incorpora información catastral
  (número de viviendas afectada por cada cambio de sección), se puede fijar
  un umbral de cambio (%) para rechazar aquellos cambios que involucren a muy
  pocas viviendas.
}
\examples{

\dontrun{
  library(medear)
  data("poblacion")
  data("cambios_seccion")
  data("cartografia")
  uniones <- une_secciones(
    cambios       = cambios_seccion,
    cartografia   = cartografia,
    years         = 2006:2016,
    poblacion     = poblacion,
    catastro      = TRUE,
    umbral_cambio = 10
  )

  poblacion   <- uniones$poblacion
  cartografia <- uniones$cartografia
}

}
\seealso{
\code{\link{detecta_cambios}}, \code{\link{descarga_poblaciones}} y
  \code{\link{descarga_cartografia}}
}
