% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_cambios.R
\encoding{UTF-8}
\name{une_secciones}
\alias{une_secciones}
\title{Une los cambios del seccionado del INE}
\usage{
une_secciones(cambios, cartografia, years = 1996:2016, poblacion =
  NULL, corte_edad = 85, catastro = FALSE, umbral_vivienda = 5, distancia =
  100)
}
\arguments{
\item{cambios}{Objeto de clase \code{cambios_ine}.}

\item{cartografia}{Objeto de clase \code{\link[sp]{SpatialPolygons}}, y con
datos de clase \code{cartografia_ine}.}

\item{years}{Vector numérico de longitud >= 1 con los años para los que se
desee consultar las variaciones de seccionado. El año 2011 debe figurar
dentro del vector, cuyo rango debe ser continuo (sin saltos de más de un
año).}

\item{poblacion}{Objeto de clase \code{poblaciones_ine}. Argumento opcional a
proporcionar en caso de querer agregar las poblaciones.}

\item{corte_edad}{Numérico: punto de corte para los grupos de edad (85 o
100). Argumento opcional en caso de proporcionar datos de poblaciones.}

\item{catastro}{Lógico: ¿Debe aplicarse el filtro de información catastral?
Por defecto \code{catastro = FALSE}.}

\item{umbral_vivienda}{Numérico: porcentaje de viviendas afectadas en el
cambio de sección. Solo se utiliza si \code{catastro = TRUE}. Por defecto
se fija al 5 \%.}

\item{distancia}{Numérico: Máxima distancia (en metros) de separación entre
secciones. Por defecto se fija en 100 metros. En algunos casos
(principalmente en ciudades donde no haya problemas con pedanías que
compartan nombres de vía con el núcleo urbano principal) puede ser
conveniente aumentar este parámetro.}
}
\value{
El resultado devuelto varía en función de si se proporcionan datos de
  poblaciones o no. Si no se proporcionan se devuelve un objeto de clase
  \code{cartografia_ine} y \code{\link[sp]{SpatialPolygonsDataFrame}} con la
  cartografía, donde cada fila es una sección censal y que cuenta con 9
  columnas: \describe{\item{seccion}{Cadena de 10 caracteres con el código de
  sección censal (incluye provincia, municipio y distrito).}
  \item{CUMUN}{Cadena de 5 caracteres con el código del municipio (incluye
  provincia).} \item{CCA}{Cadena de 2 caracteres con el código de comunidad
  autónoma.} \item{NPRO}{Nombre de la provincia.} \item{NCA}{Nombre de la
  comunidad autónoma.} \item{NMUN}{Nombre del municipio.}
  \item{geometry}{Columna de tipo lista con la geometría asociada a cada
  sección censal.} \item{cluster_id}{Código de identificación del cluster de
  uniones.} \item{sc_unida}{Código de las secciones unidas.}}

  En caso de proporcionan poblaciones, se devuelve una lista de longitud
  igual a dos, donde el primer elemento es la cartografía descrita
  anteriormente y el segundo elemento de la lista es un objeto de clase
  \code{poblaciones_ine} donde las filas representan las distintas secciones
  censales. Las tres primeras columnas son: \describe{\item{seccion}{Código
  de la sección censal.} \item{sexo}{Sexo de la población (0 = masculino; 1 =
  femenino).} \item{year}{Año de referencia.}} El resto de columnas
  representan los distintos grupos de edad, tras realizar el corte en los
  grupos de edad (85 0 100).
}
\description{
Une los cambios del seccionado del INE (tomando como referencia
  la cartografía INE 2011), adaptando a su vez las poblaciones por sexo año y
  sección censal. Si el archivo de cambios incorpora información catastral
  (número de viviendas afectada por cada cambio de sección), se puede fijar
  un umbral de cambio (\%) para rechazar aquellos cambios que involucren a
  muy pocas viviendas.
}
\details{
La función trabaja con la siguiente dinámica:

  \itemize{ \item Filtrado del archivo de cambios según el rango de años
  escogido. \item Para cada cambio, si ambas secciones existen en la
  cartografía proporcionada se calculan las distancias (en metros) entre
  ellas. \item Si se ha decidido utilizar el filtro de catastro, se calcula
  el porcentaje de viviendas implicadas en cada cambio (opción no disponible
  para Euskadi y Navarra), y se procede al filtrado del archivo de cambios
  según el umbral escogido en la llamada a la función, asegurando siempre la
  presencia de aquellos cambios que implequen a secciones que no existan en
  la cartografía proporcionada, y restringiendo el filtrado a secciones que
  disten menos de 100 metros entre sí. \item  En caso de no utilizar el
  filtro de catastro, se filtra el archivo de cambios asegurando siempre la
  presencia de aquellos cambios que implequen a secciones que no existan en
  la cartografía proporcionada, y restringiendo el filtrado a secciones que
  disten menos de 100 metros entre sí. \item Una vez que se dispone del
  archivo de cambios definitivo, se crean las agrupaciones de secciones, y se
  realiza la unión de las mismas en la cartografía. \item Si se proporciona
  un archivo de poblaciones, se agrega la población empleando las mismas
  agrupaciones de secciones del punto previo.}

  No obstante, y dado que la función asume que el callejero, el archivo de
  poblaciones y la cartografía están libres de errores. Como puede
  imaginarse, esto no es así, de modo que la función puede comportarse de
  forma inestable en dos supuestos:

  \enumerate{ \item Cuando se quiera unir cambios no solo en la cartografía
  sino también en los datos de población, puede aparecer un comportamiento
  inestable de la funció, debido a divergencias existentes en la información
  contenida en los datos de población y en los trameros (que es desde donde
  se crea el listado de cambios de sección), a pesar de que en ambos casos la
  fuente de información es el propio INE.

  Lo anterior se traduce en que, para determinadas consultas, el número de
  secciones contenidas en los datos de cartografía y poblaciones no será el
  mismo. Cuando esto pase (si pasa) la función devolverá un aviso, indicando
  qué secciones se ven afectadas y en qué años, de forma que el usuario pueda
  tratar de solucionarlo por su cuenta, aunque no hay una solución perfecta.

  Las dos soluciones más efectivas (aunque son soluciones \emph{ad hoc} y
  recae en el usuario encontrar la más apropiada para su consulta concreta)
  que se han encontrado son:

  \itemize{ \item modificar los criterios temporales de la consulta,
  ampliando o reduciendo el marco temporal (p. ej., pasar de un período
  2001:2015 a 1996:2015 o 2002:2014); \item consultar las secciones
  problemáticas (accesibles mediante la consulta \code{attr(objeto_devuelto,
  "sc_not_in_cartografia")}) en los datos de población y, en base al archivo
  de cambios de sección, decidir con qué sección se debería realizar la
  unión.}

  \item Por otra parte, es posible encontrar vías que aparecen literalmente
  "de la nada", especialmente en barrios de nueva creación o gran expansión.
  El proceso de detección de cambios de sección (función
  \code{\link{detecta_cambios}}) compara las secciones a las que se asigna
  cada tramo del callejero de 2011, con las secciones a las que se asignan
  esos mismos tramos (u otros pero contengan portales asociados a los tramos
  previos) en los callejeros del resto de años.

  No obstante, esto plantea un problema en la detección de cambios al
  considerar la aparición de vías completamente nuevas, puesto que la
  comparación 2011-otros años no es posible. En esos casos, y siempre que no
  haya uniones adicionales que resuelvan el problema por sí solo, el archivo
  de poblaciones tras la unión contendrá valores iguales a uno en todas las
  categorías de edad para los años anteriores a la creación de la vía.
  Nuevamente, cuando esto pase (si pasa) la función devolverá un aviso,
  indicando qué secciones se ven afectadas y en qué años, de forma que el
  usuario pueda tratar de solucionarlo por su cuenta.

  La solución a este problema es similar al lo anteriormente expuesto: por un
  lado se puede variar el rango de años, y por otro tratar de solucionarlo
  manualmente consultando el archivo de cambios de sección y el de
  poblaciones, buscando las secciones que devuelva la consulta
  \code{attr(resultado, "pob_igual_uno")}.}

  En el apartado de ejemplos se desarrollarán los abordajes a estos
  problemas, con un tratamiento más extenso en la viñeta de unión de
  seccionado (aún por elaborar).
}
\examples{

\dontrun{
  # En este ejemplo se trabaja con la ciudad de Sevilla (código ine: 41091)
  # en los años 2004-2015

  library(medear)
  data("poblacion")
  data("cambios_seccion")
  data("cartografia")
  cambios_se     <- cambios_seccion[substr(cambios_seccion$sc_ref, 1, 5) == "41091", ]
  cartografia_se <- cartografia[cartografia$CUMUN == "41091", ]
  poblacion_se   <- poblacion[substr(poblacion$seccion, 1, 5) == "41091", ]

  ##########################################################################
  ## Ejemplo sin utilizar el filtro de catastro                           ##
  ##########################################################################

  union_sin_cat <- une_secciones(
    cambios         = cambios_se,
    cartografia     = cartografia_se,
    years           = 2004:2015,
    poblacion       = poblacion_se,
    catastro        = FALSE,
    distancia       = 100
  )

  nrow(union_sin_cat$cartografia) # 402 secciones
  length(unique(union_sin_cat$poblacion$seccion)) # 408 secciones
  round(nrow(union_sin_cat$cartografia) / nrow(cartografia_se) * 100) #
  Conserva el 76 \\\% de secciones

  # La función avisa acerca de divergencias en el seccionado entre
  cartografía y el archivo de poblaciones. # Las secciones afectadas son
  accesibles mediante la siguiente consulta:
  attributes(union_con_cat)$sc_not_in_cartografia # 12 SC problemáticas

  # En este caso, se resolverá la primera incidencia (SC 4109102089).
  sc_problematica <- attributes(union_con_cat)$sc_not_in_cartografia[1]

  ## POR COMPLETAR ##


  ##########################################################################
  ## Ejemplo utilizando el filtro de catastro                             ##
  ##########################################################################

  union_con_cat <- une_secciones(
    cambios         = cambios_se,
    cartografia     = cartografia_se,
    years           = 2004:2015,
    poblacion       = poblacion_se,
    catastro        = TRUE,
    umbral_vivienda = 5,
    distancia       = 100
  )

  nrow(union_con_cat$cartografia) # 466 secciones
  length(unique(union_con_cat$poblacion$seccion)) # 478 secciones
  round(nrow(union_con_cat$cartografia) / nrow(cartografia_se) * 100) # Conserva el 88 \\\% de secciones

  ## POR COMPLETAR ##

}

}
\seealso{
\code{\link{detecta_cambios}}, \code{\link{descarga_poblaciones}} y
  \code{\link{descarga_cartografia}}
}
