% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_cambios.R
\encoding{UTF-8}
\name{une_secciones}
\alias{une_secciones}
\title{Une los cambios del seccionado del INE}
\usage{
une_secciones(cambios, cartografia, poblacion = NULL,
mortalidad = NULL, years = 1996:2015, corte_edad = 85, catastro = FALSE,
umbral_vivienda = 5, distancia = 100, modo = c("auto", "manual"),
sc1 = NULL, sc2 = NULL)
}
\arguments{
\item{cambios}{Objeto de clase \code{cambios_ine}.}

\item{cartografia}{Objeto de clase \code{\link[sp]{SpatialPolygons}}, y con
datos de clase \code{cartografia_ine}.}

\item{poblacion}{Objeto de clase \code{poblaciones_ine}. Argumento opcional a
proporcionar en caso de querer agregar las poblaciones.}

\item{mortalidad}{Objeto con los registros de mortalidad geocodificados. Este
 es un argumento opcional a proporcionar en caso de querer agregar la
mortalidad por sección censal. Se requiere que los datos tengan, como
mínimo, las variables sexo, año de defunción, edad, causa de defunción y
el par de coordenadas (longitud y latitud), y que tengan exactamente los
siguientes nombres: 'sexo', 'year_defuncion', 'edad', 'causa_defuncion',
'lng', y 'lat', respectivamente.}

\item{years}{Vector numérico de longitud >= 1 con los años para los que se
desee consultar las variaciones de seccionado. El año 2011 debe figurar
dentro del vector, cuyo rango debe ser continuo (sin saltos de más de un
año).}

\item{corte_edad}{Numérico: punto de corte para los grupos de edad (85 o
100). Argumento opcional en caso de proporcionar datos de poblaciones.}

\item{catastro}{Lógico: ¿Debe aplicarse el filtro de información catastral?
Por defecto \code{catastro = FALSE}.}

\item{umbral_vivienda}{Numérico: porcentaje de viviendas afectadas en el
cambio de sección. Solo se utiliza si \code{catastro = TRUE}. Por defecto
se fija al 5 \%.}

\item{distancia}{Numérico: Máxima distancia (en metros) de separación entre
secciones. Por defecto se fija en 100 metros. En algunos casos
(principalmente en ciudades donde no haya problemas con pedanías que
compartan nombres de vía con el núcleo urbano principal) puede ser
conveniente aumentar este parámetro.}

\item{modo}{Carácter: Modalidad de la unión (\code{auto} o \code{manual}), por
defecto \code{auto}. Si se escoge el modo \code{auto} la función detectará
los cambios que conllevan una unión de secciones según los parámetros que
el usuario haya elegido; si se escoge el modo \code{manual}, además de
realizar todo el proceso que involucra el otro modo, se permite al usuario
incluir uniones que el modo \code{auto} puede haber pasado por alto,
realizando uniones de seccionado 1-1 mediante los argumentos \code{sc1} y
\code{sc2}.}

\item{sc1}{Vector de caracteres: Primer bloque de secciones a unir
manualmente, vacío por defecto. Si se proporciona debe tener la misma
longitud que \code{sc2}.}

\item{sc2}{Vector de caracteres: Segundo bloque de secciones a unir
manualmente, vacío por defecto. Si se proporciona debe tener la misma
longitud que \code{sc1}.}
}
\value{
El resultado devuelto varía en función de si se proporcionan datos de
  poblaciones o mortalidad. Si no se proporciona ninguno se devuelve un
  objeto de clase \code{cartografia_ine} y \code{\link[sp]{SpatialPolygons}}
  con la cartografía, donde cada fila es una sección censal y que cuenta con
  9 columnas: \describe{\item{seccion}{Cadena de 10 caracteres con el código
  de sección censal (incluye provincia, municipio y distrito).}
  \item{CUMUN}{Cadena de 5 caracteres con el código del municipio (incluye
  provincia).} \item{CCA}{Cadena de 2 caracteres con el código de comunidad
  autónoma.} \item{NPRO}{Nombre de la provincia.} \item{NCA}{Nombre de la
  comunidad autónoma.} \item{NMUN}{Nombre del municipio.}
  \item{geometry}{Columna de tipo lista con la geometría asociada a cada
  sección censal.} \item{cluster_id}{Código de identificación del cluster de
  uniones.} \item{revision_manual}{Indica si debe revisarse esa unión.}}

  En caso de proporcionar poblaciones, se devuelve una lista de longitud
  igual a dos, donde el primer elemento es la cartografía descrita
  anteriormente y el segundo elemento de la lista es un objeto de clase
  \code{array} con cuatro dimensiones: año de defunción, sexo (0 =
  masculino; 1 = femenino), grupo de edad (según corte establecido) y
  sección censal.

  En caso de proporcionar datos de mortalidad geocodificada, se devuelve una
  lista de longitud igual a tres, donde los primeros elementos son los
  anteriormente descritos y el tercer elemento es un objeto de clase
  \code{array} con cinco dimensiones: año de defunción, sexo, grupo de edad
  (según corte establecido), sección censal y causa de muerte.
}
\description{
Une los cambios del seccionado del INE (tomando como referencia
  la cartografía INE 2011), adaptando a su vez las poblaciones por sexo,
  año, grupo de edad y sección censal o la mortalidad por sexo, año de
  defunción, grupo de edad, sección censal y causa de defunción (siempre que
  se desee, pues son argumentos opcionales). Si el archivo de cambios
  incorpora información catastral (número de viviendas afectada por cada
  cambio de sección), se puede fijar un umbral de cambio (\%) para rechazar
  aquellos cambios que involucren a muy pocas viviendas.
}
\details{
La función trabaja con la siguiente dinámica:

  \itemize{ \item Filtrado del archivo de cambios según el rango de años
  escogido. \item Para cada cambio, si ambas secciones existen en la
  cartografía proporcionada se calculan las distancias (en metros) entre
  ellas. \item Si se ha decidido utilizar el filtro de catastro, se calcula
  el porcentaje de viviendas implicadas en cada cambio (opción no disponible
  para Euskadi y Navarra), y se procede al filtrado del archivo de cambios
  según el umbral escogido en la llamada a la función, asegurando siempre la
  presencia de aquellos cambios que implequen a secciones que no existan en
  la cartografía proporcionada, y restringiendo el filtrado a secciones que
  disten menos de 100 metros entre sí. \item  En caso de no utilizar el
  filtro de catastro, se filtra el archivo de cambios asegurando siempre la
  presencia de aquellos cambios que implequen a secciones que no existan en
  la cartografía proporcionada, y restringiendo el filtrado a secciones que
  disten menos de 100 metros entre sí. \item Una vez que se dispone del
  archivo de cambios definitivo, se crean las agrupaciones de secciones, y se
  realiza la unión de las mismas en la cartografía. \item Si se proporciona
  un archivo de poblaciones, se agrega la población empleando las mismas
  agrupaciones de secciones. \item Si se proporciona un archivo de
  mortalidad geocodificada, se agregan las defunciones empleando las mismas
  agrupaciones de secciones.}

  No obstante, y dado que la función asume que el callejero, el archivo de
  poblaciones y la cartografía están libres de errores. Como puede
  imaginarse, esto no es así, de modo que la función puede comportarse de
  forma inestable en dos supuestos:

  \enumerate{ \item Cuando se quiera unir cambios no solo en la cartografía
  sino también en los datos de población, puede aparecer un comportamiento
  inestable de la función, debido a divergencias existentes en la información
  contenida en los datos de población y en los trameros (que es desde donde
  se crea el listado de cambios de sección), a pesar de que en ambos casos la
  fuente de información es el propio INE.

  Lo anterior se traduce en que, para determinadas consultas, el número de
  secciones contenidas en los datos de cartografía y poblaciones no será el
  mismo. Cuando esto pase (si pasa) la función devolverá un aviso, indicando
  qué secciones se ven afectadas y en qué años, de forma que el usuario pueda
  tratar de solucionarlo por su cuenta, aunque no hay una solución perfecta.

  Las dos soluciones más efectivas (aunque son soluciones \emph{ad hoc} y
  recae en el usuario encontrar la más apropiada para su consulta concreta)
  que se han encontrado son:

  \itemize{ \item modificar los criterios temporales de la consulta,
  ampliando o reduciendo el marco temporal (p. ej., pasar de un período
  2001:2015 a 1996:2015 o 2002:2014); \item consultar las secciones
  problemáticas (accesibles mediante la consulta \code{attr(objeto_devuelto,
  "sc_not_in_cartografia")}) en los datos de población y, en base al archivo
  de cambios de sección, decidir con qué sección se debería realizar la
  unión.}

  \item Por otra parte, es posible encontrar vías que aparecen literalmente
  "de la nada", especialmente en barrios de nueva creación o gran expansión.
  El proceso de detección de cambios de sección (función
  \code{\link{detecta_cambios}}) compara las secciones a las que se asigna
  cada tramo del callejero de 2011, con las secciones a las que se asignan
  esos mismos tramos (u otros pero contengan portales asociados a los tramos
  previos) en los callejeros del resto de años.

  No obstante, esto plantea un problema en la detección de cambios al
  considerar la aparición de vías completamente nuevas, puesto que la
  comparación 2011-otros años no es posible. En esos casos, y siempre que no
  haya uniones adicionales que resuelvan el problema por sí solo, el archivo
  de poblaciones tras la unión contendrá valores iguales a uno en todas las
  categorías de edad para los años anteriores a la creación de la vía.
  Nuevamente, cuando esto pase (si pasa) la función devolverá un aviso,
  indicando qué secciones se ven afectadas y en qué años, de forma que el
  usuario pueda tratar de solucionarlo por su cuenta.

  La solución a este problema es similar al lo anteriormente expuesto: por un
  lado se puede variar el rango de años, y por otro tratar de solucionarlo
  manualmente consultando el archivo de cambios de sección y el de
  poblaciones, buscando las secciones que devuelva la consulta
  \code{attr(resultado, "pob_igual_uno")}.}

  En el apartado de ejemplos se desarrollarán los abordajes a estos
  problemas, con un tratamiento más extenso en la viñeta de unión de
  seccionado (aún por elaborar).
}
\examples{

\dontrun{
  # En este ejemplo se trabaja con la ciudad de Córdoba (código ine: 14021)
  # en los años 2004-2015, sin usar códigos postales.

  library(medear)
  data("poblacion")
  data("cambios_seccion")
  data("cartografia")
  cartografia_co <- cartografia[cartografia$CUMUN == "14021", ]
  poblacion_co   <- poblacion[substr(poblacion$seccion, 1, 5) == "14021", ]
  cambios_co     <- cambios_seccion[
    substr(cambios_seccion$sc_ref, 1, 5) == "14021" & cambios_seccion$codigo_postal == FALSE,
  ]

  ##########################################################################
  ## Ejemplo sin utilizar el filtro de catastro                           ##
  ##########################################################################

  union_sin_cat <- une_secciones(
    cambios     = cambios_co,
    cartografia = cartografia_co,
    years       = 2004:2015,
    poblacion   = poblacion_co,
    catastro    = FALSE,
    distancia   = 100
  )

  nrow(union_sin_cat$cartografia) # 215 secciones
  round(nrow(union_sin_cat$cartografia) / nrow(cartografia_co) * 100) #
  Conserva el 87 \\\% de secciones

  # La función avisa acerca de divergencias en el seccionado entre
  cartografía y el archivo de poblaciones. # Las secciones afectadas son
  accesibles mediante la siguiente consulta:
  attributes(union_sin_cat)$pob_igual_uno # 1 SC problemática

  # En este caso, se resolverá la incidencia (SC 1402106047).
  sc_problematica <- attributes(union_sin_cat)$pob_igual_uno

  # Hay que revisar el archivo de cambios para comprobar si esta SC está
  # involucrada en algún otro cambio de sección:
  cambios_res <- attributes(union_sin_cat)$cambios
  cambios_res[cambios_res$sc_ref == sc_problematica]
  cambios_res[cambios_res$sc_new == sc_problematica]

  # La SC debería unirse con la SC 1402106024, aunque ambas están a 104.9
  # metros. Comprobemos esa otra sección:
  cambios_res[cambios_res$sc_ref == "1402106024"]
  cambios_res[cambios_res$sc_new == "1402106024"]

  # La SC 1402106024 se une con la SC 1402106050 (no está en la cartografía
  # de 2001) y con la 1402106038. Al representar estas SC (1402106047,
  # y la unión 1402106024-1402106038), vemos que colindan, así que se puede
  # realizar la unión manual:
  plot(union_sin_cat$cartografia[union_sin_cat$cartografia$seccion \%in\%
    c("1402106047", "1402106024"), ])
  union_sin_cat <- une_secciones(
    cambios     = cambios_co,
    cartografia = cartografia_co,
    years       = 2004:2015,
    poblacion   = poblacion_co,
    catastro    = FALSE,
    distancia   = 100,
    modo        = "manual",
    sc1         = "1402106047",
    sc2         = "1402106024"
  )

  # El aviso ha desaparecido, aunque todavía se mantiene un aviso sobre
  # secciones conformadas por varios polígonos. Vamos a ver qué secciones
  # forman esta unión:
  union_sin_cat$cartografia@data[union_sin_cat$cartografia$seccion == "1402101006", ]

  # La sección involucrada es 1402106010. Al representar ambas secciones
  # vemos que sí se tocan, aunque solo en una esquina. A pesar del aviso,
  # damos por buena la unión.
  plot(union_sin_cat$cartografia[union_sin_cat$cartografia$seccion == "1402101006", ])

  ##########################################################################
  ## Ejemplo utilizando el filtro de catastro                             ##
  ##########################################################################

  union_con_cat <- une_secciones(
    cambios         = cambios_co,
    cartografia     = cartografia_co,
    years           = 2004:2015,
    poblacion       = poblacion_co,
    catastro        = TRUE,
    umbral_vivienda = 5,
    distancia       = 100
  )

  nrow(union_con_cat$cartografia) # 223 secciones
  round(nrow(union_con_cat$cartografia) / nrow(cartografia_co) * 100)
  # Conserva el 91 \\\% de secciones

  # Al utilizar información de catastro se ha aumentado de un 88 \\\% a un
  # 91 \\\%, recuperando 8 secciones. Surgen los mismos avisos que antes, así
  # que procedemos de la misma forma, tras lo cual se da por cerrada la unión:
  union_con_cat <- une_secciones(
    cambios         = cambios_co,
    cartografia     = cartografia_co,
    years           = 2004:2015,
    poblacion       = poblacion_co,
    catastro        = TRUE,
    umbral_vivienda = 5,
    distancia       = 100,
    modo            = "manual",
    sc1             = "1402106047",
    sc2             = "1402106024"
  )

}

}
\seealso{
\code{\link{detecta_cambios}}, \code{\link{descarga_poblaciones}} y
  \code{\link{descarga_cartografia}}
}
